isaacgymenvs copy/PAIR.py
# Required libraries (you might need more based on your implementation details)
import numpy as np
import tensorflow as tf

# ===============================
# 1. Environment Generator (Adversary)
# ===============================

class EnvironmentTemplate:
    def __init__(self, width, height):
        self.width = width
        self.height = height
        self.grid = np.zeros((width, height))

class Adversary:
    def __init__(self):
        # Initialize adversary's parameters/policy
        pass
    
    def generate_environment(self):
        # Generate a new environment based on current policy
        env = EnvironmentTemplate(10, 10)
        # Add goals, obstacles, etc. to the env.grid
        return env
    
    def update_adversary(self, regret):
        # Update the adversary's policy based on the regret
        pass

# ===============================
# 2. Agent Trainer (Protagonist & Antagonist)
# ===============================

class Agent:
    def __init__(self, agent_type):
        self.agent_type = agent_type  # 'protagonist' or 'antagonist'
        # Initialize agent's parameters/policy
    
    def train_agent(self, environment):
        # Train the agent in the provided environment
        pass
    
    def evaluate_agent(self, environment):
        # Evaluate the agent's performance in the environment
        # Return performance score
        return score

# ===============================
# 3. Regret Calculator
# ===============================

def calculate_regret(protagonist_score, antagonist_score):
    # Calculate the regret based on agent performances
    regret = antagonist_score - protagonist_score
    return regret

# ===============================
# 4. Main Training Loop
# ===============================

def main_training_loop(epochs):
    # Initialize agents and adversary
    protagonist = Agent("protagonist")
    antagonist = Agent("antagonist")
    adversary = Adversary()
    
    for epoch in range(epochs):
        # Generate environment
        environment = adversary.generate_environment()
        
        # Train and evaluate protagonist
        protagonist.train_agent(environment)
        protagonist_score = protagonist.evaluate_agent(environment)
        
        # Evaluate antagonist (assuming antagonist doesn't train but only evaluates)
        antagonist_score = antagonist.evaluate_agent(environment)
        
        # Calculate regret
        regret = calculate_regret(protagonist_score, antagonist_score)
        
        # Update adversary
        adversary.update_adversary(regret)

# ===============================
# 5. Visualization & Debugging Tools
# ===============================

def visualize_environment(environment):
    # Visualize the environment grid (e.g., using matplotlib or any other library)
    pass

def visualize_agent_path(agent, environment):
    # Visualize the path taken by the agent in the environment
    pass

# ===============================
# 6. Zero-Shot Evaluation
# ===============================

def zero_shot_evaluation(agent, test_environments):
    # Evaluate the agent's performance in unseen environments
    for environment in test_environments:
        score = agent.evaluate_agent(environment)
        # Store/Print the score

# ===============================
# 7. Optional: Population-Based Training (PBT) Extension
# ===============================

# You can create a pool of agents and adversaries and select from them in the main_training_loop.

# ===============================
# Run the training
# ===============================

if __name__ == "__main__":
    main_training_loop(epochs=1000)
