isaacgymenvs copy/requirements.txt
absl-py==2.0.0
antlr4-python3-runtime==4.9.3
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
cloudpickle==3.0.0
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
docker-pycreds==0.4.0
filelock==3.13.0
fonttools==4.43.1
freetype-py==2.4.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.3
google-auth-oauthlib==1.0.0
grpcio==1.59.0
gym==0.23.1
gym-notices==0.0.8
hydra-core==1.3.2
idna==3.4
imageio==2.31.6
importlib-metadata==6.8.0
importlib-resources==6.1.0
# Editable install with no version control (isaacgym==1.0rc4)
-e /home/linjiw/Downloads/IsaacGym_Preview_4_Package/isaacgym/python
-e git+https://github.com/NVIDIA-Omniverse/IsaacGymEnvs.git@3f847a7f20feddc1e6ccd988765cc1a24e72ac7c#egg=isaacgymenvs
Jinja2==3.1.2
kiwisolver==1.4.5
lxml==4.9.3
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.4.0
mpmath==1.3.0
networkx==2.2
ninja==1.11.1.1
numpy==1.19.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.8.1.78
packaging==23.2
pathtools==0.1.2
Pillow==10.0.1
promise==2.3
protobuf==3.20.3
psutil==5.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycollada==0.6
pyglet==2.0.9
PyOpenGL==3.1.0
pyparsing==3.1.1
pyrender==0.1.45
pysdf==0.1.9
python-dateutil==2.8.2
PyVirtualDisplay==3.0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rl_games==1.6.1
rsa==4.9
scipy==1.10.1
sentry-sdk==1.32.0
setproctitle==1.3.3
shortuuid==1.0.11
six==1.16.0
smmap==5.0.1
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.3.0
torch==2.1.0
torchvision==0.16.0
trimesh==3.23.5
triton==2.1.0
typing_extensions==4.8.0
urdfpy==0.0.22
urllib3==2.0.7
wandb==0.12.21
warp-lang==0.10.1
Werkzeug==3.0.1
zipp==3.17.0
